[WebCodecs] Fix AV1 color space settings.
This seems to have broken somewhere along the line. The type
deduced by the compiler is something not-int when libaom is
expecting int. Not sure why it isn't showing up on bots, but
it broke the full-cycle-encoding test locally for me.

R=eugene

Bug: 1377842
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4226398
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101866}

+++ media/video/av1_video_encoder.cc
@@ -598,19 +598,20 @@ void Av1VideoEncoder::UpdateEncoderColorSpace() {
   auto aom_cs = VideoColorSpace::FromGfxColorSpace(last_frame_color_space_);
   if (aom_cs.primaries != VideoColorSpace::PrimaryID::INVALID) {
     auto status = aom_codec_control(codec_.get(), AV1E_SET_COLOR_PRIMARIES,
-                                    aom_cs.primaries);
+                                    static_cast<int>(aom_cs.primaries));
     if (status != AOM_CODEC_OK)
       LogAomErrorMessage(codec_.get(), "Failed to set color primaries", status);
   }
   if (aom_cs.transfer != VideoColorSpace::TransferID::INVALID) {
-    auto status = aom_codec_control(
-        codec_.get(), AV1E_SET_TRANSFER_CHARACTERISTICS, aom_cs.transfer);
+    auto status =
+        aom_codec_control(codec_.get(), AV1E_SET_TRANSFER_CHARACTERISTICS,
+                          static_cast<int>(aom_cs.transfer));
     if (status != AOM_CODEC_OK)
       LogAomErrorMessage(codec_.get(), "Failed to set color transfer", status);
   }
   if (aom_cs.matrix != VideoColorSpace::MatrixID::INVALID) {
     auto status = aom_codec_control(codec_.get(), AV1E_SET_MATRIX_COEFFICIENTS,
-                                    aom_cs.matrix);
+                                    static_cast<int>(aom_cs.matrix));
     if (status != AOM_CODEC_OK)
       LogAomErrorMessage(codec_.get(), "Failed to set color transfer", status);
   }
